Sidebar now shows a users google photo and name
